+++ packages/open-next/src/adapters/server-adapter.ts
@@ -11,12 +11,13 @@ import type {
 import NextServer from "next/dist/server/next-server.js";
 //@ts-ignore
 import { getMaybePagePath } from "next/dist/server/require.js";
-import { loadConfig, setNodeEnv } from "./util.js";
+import { generateUniqueId, loadConfig, setNodeEnv } from "./util.js";
 import { isBinaryContentType } from "./binary.js";
 import { debug } from "./logger.js";
 import type { PublicFiles } from "../build.js";
 import { convertFrom, convertTo } from "./event-mapper.js";
 import { overrideReact } from "./require-hooks.js";
+import { WarmerEvent, WarmerResponse } from "./warmer-function.js";
 
 setNodeEnv();
 setNextjsServerWorkingDirectory();
@@ -28,7 +29,8 @@ const publicAssets = loadPublicAssets();
 initializeNextjsRequireHooks(config);
 debug({ nextDir });
 
-// Create a NextServer
+// Generate a 6 letter unique server ID
+const serverId = `server-${generateUniqueId()}`;
 const requestHandler = new NextServer.default({
   hostname: "localhost",
   port: Number(process.env.PORT) || 3000,
@@ -45,8 +47,17 @@ const requestHandler = new NextServer.default({
 /////////////
 
 export async function handler(
-  event: APIGatewayProxyEventV2 | CloudFrontRequestEvent | APIGatewayProxyEvent
+  event:
+    | APIGatewayProxyEventV2
+    | CloudFrontRequestEvent
+    | APIGatewayProxyEvent
+    | WarmerEvent
 ) {
+  // Handler warmer
+  if ("type" in event) {
+    return formatWarmerResponse(event);
+  }
+
   debug("event", event);
 
   // Parse Lambda event and create Next.js request
@@ -187,4 +198,13 @@ function formatAPIGatewayFailoverResponse() {
 
 function formatCloudFrontFailoverResponse(event: CloudFrontRequestEvent) {
   return event.Records[0].cf.request;
+}
+
+function formatWarmerResponse(event: WarmerEvent) {
+  console.log(event);
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve({ serverId } satisfies WarmerResponse);
+    }, event.delay);
+  });
 }